rename files part 2
