+++ jsapp/js/components/map/MapSettings.tsx
